>> ripped cct's from piazza, began placer logic.
-- found and modified bug in intialize_system (added a zero size blck-nets)

+++ source/configholder.h
@@ -7,16 +7,16 @@ class Configholder {
 
 	private: 
 		vector< vector<int> > blck_to_nets;
-		vector< vector<int> > ref_blcks;
+		vector< vector<float> > ref_blcks;
 
 	public:
 		Configholder(){};
 
 
 		void add_blck_to_net(vector<int> blck_to_net);
-		void add_ref_blck(vector<int> ref_blck);
+		void add_ref_blck(vector<float> ref_blck);
 		vector< vector<int> > get_blck_to_nets();
-		vector< vector<int> > get_ref_blcks();
+		vector< vector<float> > get_ref_blcks();
 		int get_grid_size();
 		void display_config();
